refactor(bridge): extract view-once sanitization
